+++ Source/SwiftLintFramework/Protocols/Rule.swift
@@ -9,26 +9,47 @@
 import SourceKittenFramework
 
 public protocol Rule {
+    init() // Rules need to be able to be initialized with default values
     static var description: RuleDescription { get }
     func validateFile(file: File) -> [StyleViolation]
 }
 
 extension Rule {
     func isEqualTo(rule: Rule) -> Bool {
-        return self.dynamicType.description == rule.dynamicType.description
+        switch (self, rule) {
+        case (let rule1 as ConfigurableRule, let rule2 as ConfigurableRule):
+            return rule1.isEqualTo(rule2)
+        default:
+            return self.dynamicType.description == rule.dynamicType.description
+        }
     }
 }
 
-public protocol ParameterizedRule: Rule {
+public protocol ConfigurableRule: Rule {
+    init?(config: AnyObject)
+    func isEqualTo(rule: ConfigurableRule) -> Bool
+}
+
+public protocol ParameterizedRule: ConfigurableRule {
     typealias ParameterType: Equatable
     init(parameters: [RuleParameter<ParameterType>])
     var parameters: [RuleParameter<ParameterType>] { get }
 }
 
+// Default implementation for ConfigurableRule conformance
 extension ParameterizedRule {
-    func isEqualTo(rule: Self) -> Bool {
-        return (self.dynamicType.description == rule.dynamicType.description) &&
-               (self.parameters == rule.parameters)
+    public init?(config: AnyObject) {
+        guard let array = [ParameterType].arrayOf(config) else {
+            return nil
+        }
+        self.init(parameters: RuleParameter<ParameterType>.ruleParametersFromArray(array))
+    }
+
+    public func isEqualTo(rule: ConfigurableRule) -> Bool {
+        if let rule = rule as? Self {
+            return parameters == rule.parameters
+        }
+        return false
     }
 }
 
@@ -38,7 +59,7 @@ public protocol CorrectableRule: Rule {
 
 // MARK: - == Implementations
 
-func == (lhs: [Rule], rhs: [Rule]) -> Bool {
+public func == (lhs: [Rule], rhs: [Rule]) -> Bool {
     if lhs.count == rhs.count {
         return zip(lhs, rhs).map { $0.isEqualTo($1) }.reduce(true) { $0 && $1 }
     }