+++ Source/SwiftLintFramework/Rules/CommaRule.swift
@@ -10,6 +10,9 @@ import Foundation
 import SourceKittenFramework
 
 public struct CommaRule: CorrectableRule {
+
+    public init() {}
+
     public static let description = RuleDescription(
         identifier: "comma",
         name: "Comma Spacing",