+++ Source/SwiftLintFramework/Rules/StatementPositionRule.swift
@@ -10,6 +10,9 @@ import Foundation
 import SourceKittenFramework
 
 public struct StatementPositionRule: CorrectableRule {
+
+    public init() {}
+
     public static let description = RuleDescription(
         identifier: "statement_position",
         name: "Statement Position",