+++ Source/SwiftLintFramework/Rules/TrailingWhitespaceRule.swift
@@ -9,6 +9,9 @@
 import SourceKittenFramework
 
 public struct TrailingWhitespaceRule: CorrectableRule {
+
+    public init() {}
+
     public static let description = RuleDescription(
         identifier: "trailing_whitespace",
         name: "Trailing Whitespace",