+++ Source/SwiftLintFramework/Rules/TypeNameRule.swift
@@ -10,6 +10,9 @@ import SourceKittenFramework
 import SwiftXPC
 
 public struct TypeNameRule: ASTRule {
+
+    public init() {}
+
     public static let description = RuleDescription(
         identifier: "type_name",
         name: "Type Name",