+++ Source/SwiftLintFramework/Rules/VariableNameMinLengthRule.swift
@@ -9,19 +9,35 @@
 import SourceKittenFramework
 import SwiftXPC
 
-public struct VariableNameMinLengthRule: ASTRule, ParameterizedRule {
-    public init() {
-        self.init(parameters: [
-            RuleParameter(severity: .Warning, value: 3),
-            RuleParameter(severity: .Error, value: 2)
-        ])
+public struct VariableNameMinLengthRule: ASTRule, ConfigurableRule {
+
+    public init() { }
+
+    public init?(config: AnyObject) {
+        self.init()
+        if let config = [Int].arrayOf(config) where config.count > 0 {
+            warning = RuleParameter(severity: .Warning, value: config[0])
+            if config.count > 1 {
+                error = RuleParameter(severity: .Error, value: config[1])
+            }
+        } else if let config = config as? [String: AnyObject] {
+            if let warningNumber = config["warning"] as? Int {
+                warning = RuleParameter(severity: .Warning, value: warningNumber)
+            }
+            if let errorNumber = config["error"] as? Int {
+                error = RuleParameter(severity: .Error, value: errorNumber)
+            }
+            if let excluded = config["excluded"] as? [String] {
+                self.excluded = excluded
+            }
+        } else {
+            return nil
+        }
     }
 
-    public init(parameters: [RuleParameter<Int>]) {
-        self.parameters = parameters
-    }
-
-    public let parameters: [RuleParameter<Int>]
+    public var excluded = [String]()
+    private var warning = RuleParameter(severity: .Warning, value: 3)
+    private var error = RuleParameter(severity: .Error, value: 2)
 
     public static let description = RuleDescription(
         identifier: "variable_name_min_length",
@@ -42,15 +58,29 @@ public struct VariableNameMinLengthRule: ASTRule, ParameterizedRule {
     public func validateFile(file: File, kind: SwiftDeclarationKind,
                              dictionary: XPCDictionary) -> [StyleViolation] {
         return file.validateVariableName(dictionary, kind: kind).map { name, offset in
-            let charCount = name.characters.count
-            for parameter in parameters.reverse() where charCount < parameter.value {
-                return [StyleViolation(ruleDescription: self.dynamicType.description,
-                    severity: parameter.severity,
-                    location: Location(file: file, byteOffset: offset),
-                    reason: "Variable name should be \(parameter.value) characters " +
-                            "or more: currently \(charCount) characters")]
+            if !excluded.contains(name) {
+                let charCount = name.characters.count
+                for parameter in [error, warning] where charCount < parameter.value {
+                    return [StyleViolation(ruleDescription: self.dynamicType.description,
+                        severity: parameter.severity,
+                        location: Location(file: file, byteOffset: offset),
+                        reason: "Variable name should be \(parameter.value) characters " +
+                        "or more: currently \(charCount) characters")]
+                }
             }
             return []
         } ?? []
+    }
+
+    public func isEqualTo(rule: ConfigurableRule) -> Bool {
+        guard let rule = rule as? VariableNameMinLengthRule else {
+            return false
+        }
+
+        // Need to use alternate method to compare excluded due to apparent bug in
+        // the way that SwiftXPC compares [String]
+        return self.error == rule.error &&
+            self.warning == rule.warning &&
+            zip(self.excluded, rule.excluded).reduce(true) { $0 && ($1.0 == $1.1) }
     }
 }