+++ Source/SwiftLintFrameworkTests/TestHelpers.swift
@@ -12,9 +12,7 @@ import XCTest
 
 private let violationMarker = "↓"
 
-let allRuleIdentifiers = Configuration.rulesFromYAML().map {
-    $0.dynamicType.description.identifier
-}
+let allRuleIdentifiers = Array(masterRuleList.list.keys)
 
 func violations(string: String, config: Configuration = Configuration()) -> [StyleViolation] {
     let stringStrippingMarkers = string.stringByReplacingOccurrencesOfString(violationMarker,
@@ -101,5 +99,18 @@ extension XCTestCase {
 
         // corrections
         ruleDescription.corrections.forEach(assertCorrection)
+    }
+
+    func checkError<T: protocol<ErrorType, Equatable>>(error: T, closure: () throws -> () ) {
+        do {
+            try closure()
+            XCTFail("No error caught")
+        } catch let rError as T {
+            if error != rError {
+                XCTFail("Wrong error caught")
+            }
+        } catch {
+            XCTFail("Wrong error caught")
+        }
     }
 }